blog/requirements.txt
alembic             ==1.13.1
asyncpg             ==0.29.0
coverage            ==7.5.1
fastapi             ==0.109.2
gunicorn            ==21.2.0
httpx               ==0.27.0
orjson              ==3.9.13
pydantic            ==2.6.1
pydantic-settings   ==2.1.0
PyJWT               ==2.8.0
pytest              ==8.2.0
pytest-asyncio      ==0.23.6
python-dotenv       ==1.0.1
redis               ==5.0.1
SQLAlchemy          ==2.0.25
sqlalchemy[asyncio]
SQLAlchemy-Utils    ==0.41.1
uvicorn             ==0.27.0.post1
uvloop              ==0.19.0 ; sys_platform != "win32" and implementation_name == "cpython"

#fastapi Form data requires "python-multipart"
#python-multipart
